# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VSM Mobile",
        "description": "Documentation API VSM PK11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion d'un utilisateur",
                "description": "Connexion de l'utilisateur / membre VSM",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+237690000000"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|laravel_sanctum_token..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Déconnexion de l'utilisateur",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserModel": {
                "title": "Modèle Utilisateur / Membre VSM",
                "description": "Représentation JSON d'un utilisateur renvoyée à Flutter",
                "required": [
                    "id",
                    "name",
                    "phone",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Emmanuel Dika"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+237690000000"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@vsm.cm",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/avatars/user_1.jpg",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "player",
                        "enum": [
                            "admin",
                            "player",
                            "treasurer",
                            "president",
                            "coach"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrez votre jeton Sanctum sous la forme: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentification",
            "description": "Authentification"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}